Move logic out of article page
