+++ src/pages/article/[path].js
@@ -1,9 +1,6 @@
 import React from "react";
 import { faRedditAlien, faTwitter } from "@fortawesome/free-brands-svg-icons";
 import styled from "styled-components";
-import path from "path";
-import fs from "fs";
-import htmlToImage from "node-html-to-image";
 import { getAllArticles, getArticleBySlug } from "../../lib/fetchArticles";
 import PageTemplate from "../../components/templates/PageTemplate";
 import Container from "../../components/templates/Container";
@@ -12,6 +9,7 @@ import Markdown from "../../components/atoms/Markdown";
 import { ArticleService } from "../../services/ArticleService";
 import ShareButtons from "../../components/molecules/ShareButtons";
 import addUtmParams from "../../helpers/addUtmParams";
+import {generateArticlePreviewImage} from "../../lib/generateArticlePreviewImage";
 
 const ArticleMeta = styled("p")`
 	margin: 0;
@@ -82,74 +80,8 @@ export async function getStaticPaths() {
 
 export async function getStaticProps({ params }) {
 	const data = getArticleBySlug(params.path);
-	const publicPath = path.join(process.cwd(), "public");
-	const articleAssetsPath = path.join(publicPath, "article-assets", params.path);
 
-	if (!fs.existsSync(articleAssetsPath)) {
-		await fs.promises.mkdir(articleAssetsPath, { recursive: true });
-	}
-
-	const metaImagePath = path.join(publicPath, "article-assets", params.path, "meta-image.jpeg");
-
-	await htmlToImage({
-		type: "jpeg",
-		output: metaImagePath,
-		content: {
-			title: data.title,
-			author: data.author
-		},
-		html: `
-			<html>
-				<head>
-					<link rel="preconnect" href="https://fonts.googleapis.com">
-					<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-					<link href="https://fonts.googleapis.com/css2?family=Open+Sans&family=Raleway:wght@900&display=swap" rel="stylesheet">
-					<style>
-						body {
-						  width: 1920px;
-						  height: 1080px;
-						  margin: 0;
-						  padding: 50px;
-						  display: grid;
-						  align-items: center;
-						  border-bottom: 20px solid #1457B3;
-						  background-color: #F7F7F7;
-						}
-
-						#logo {
-						  width: 100px;
-						  position: fixed;
-						  top: 50px;
-						  left: 50px;
-						}
-
-						#title {
-						  font-family: 'Raleway', sans-serif;
-						  font-size: 100px;
-						  padding-left: 20px;
-						  border-left: 10px solid #1457B3;
-						}
-
-						#author {
-						  font-family: 'Open Sans', sans-serif;
-						  font-size: 50px;
-						}
-					</style>
-				</head>
-				<body>
-					<div id="container">
-						<img alt="CodeSupport Logo" id="logo" src="https://codesupport.dev/logo.png" />
-						<h1 id="title">
-							{{title}}
-						</h1>
-						<p id="author">
-							By {{author}}
-						</p>
-					</div>
-				</body>
-			</html>
-		`
-	});
+	await generateArticlePreviewImage(data, params.path);
 
 	return {
 		props: { data }
